requirements.txt
beautifulsoup4==4.11.1
numpy==1.21.2
pandas==1.2.5
python-dotenv==0.20.0
recipe_scrapers==13.31.0
requests==2.27.1
telegram==0.0.1
tensorflow==2.4.1
